+++ R/print.cgaim.R
@@ -1,14 +1,15 @@
 #' Print a cgaim object
 #'
-#' Default method to print the result from a call to \code{cgaim}.
-#'
+#' Default method to print the results from a call to \code{cgaim}. 
 #' Conveniently prints the formula, \code{beta} and \code{alpha} coefficients
 #'    as well as the residual sum of squares.
 #'    
 #' @param x A \code{cgaim} object.
 #' @param ... For compatibility with the default \code{print} method. Unused
 #'    at the moment.
-#'
+#' 
+#' @seealso The main fitting function \code{\link{cgaim}}.
+#' 
 #' @export
 print.cgaim <- function(x, ...){
   cat("Formula:\n")
@@ -25,4 +26,5 @@ print.cgaim <- function(x, ...){
   cat("\nResidual sum of squares: ")
   cat(x$rss)
   cat("\n")
+  invisible(x)
 }