+++ R/update_quadprog.R
@@ -1,4 +1,8 @@
-
+################################################################################
+#
+# Wrapper for quadprog alpha update
+#
+################################################################################
 
 update_quadprog <- function(Dmat, dvec, Cmat, bvec, qp_pars){
   res <- quadprog::solve.QP(Dmat, dvec, t(Cmat), bvec = bvec)